requirements.txt
astroid==1.6.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.10
django-crispy-forms==1.7.0
django-toolbelt==0.0.1
gunicorn==19.7.1
image==1.5.17
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.0.0
psycopg2==2.7.4
pylint==1.8.1
python-decouple==3.1
six==1.11.0
static3==0.7.0
waitress==1.1.0
whitenoise==3.3.1
wrapt==1.10.11
astroid==1.6.0
boto3==1.6.6
botocore==1.9.6
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.10
django-crispy-forms==1.7.0
django-storages==1.6.5
django-toolbelt==0.0.1
docutils==0.14
gunicorn==19.7.1
image==1.5.17
isort==4.2.15
jmespath==0.9.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.0.0
psycopg2==2.7.4
pylint==1.8.1
python-dateutil==2.6.1
python-decouple==3.1
s3transfer==0.1.13
six==1.11.0
static3==0.7.0
waitress==1.1.0
whitenoise==3.3.1
wrapt==1.10.11
